Cinema SQL Modification (Version: 0.2.0)
In this commit, I've made adjustments to the cinema.sql file. Specifically, I commented out the insert statements for projections. The modification aims to temporarily disable the insertion of projections directly into the database.

Additionally, FilmProjection Controller Update:
- Introduced a new POST method allowing the generation of projections.
- This method generates three projections per film for each hall every day for a week, starting from the day they are created.

These changes are part of ongoing development, refining the database structure and enhancing the FilmProjection Controller to automate the generation of projections for better system testing and management.